backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.12
python-dotenv==1.0.1
pydantic[email]==2.13.3
pydantic-settings==2.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
alembic==1.13.1

# Tests
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
